Reduced menu sidebar
Basic Setting swapped with Advanced Setting.

Basic Settings--> Business Functions (collapsible), Owners(collapsible), Organizations(collapsible), Business Info

Advanced Settings -->Backstage Access and Job table view

Added Calendar Page to another tab in settings

Added quick guide to top question mark button.
