+++ components/landing_page/navbar.tsx
@@ -527,6 +527,10 @@ const Navbar = () => {
     }
   };
 
+  const openQuickGuide = () => {
+    window.open("https://coda.io/@urvashi-batra/prioriwise-getting-started-guide", "_blank");
+  };
+
   const handleStartTourClick = () => {
     if (pathname === "/jobs") {
       const timestamp = Date.now();
@@ -663,6 +667,12 @@ const Navbar = () => {
                   Get familiar with our interface by taking a guided tour of the
                   main features.
                 </p>
+                <Button
+                className="w-full bg-[#f05523] hover:bg-[#f05523]/90 text-white"
+                onClick={openQuickGuide}
+              >
+                Get Quick Guide
+              </Button>
                 <Button
                   className="w-full bg-[#f05523] hover:bg-[#f05523]/90 text-white"
                   onClick={handleStartTourClick}
